+++ CHANGELOG.rst
@@ -1,4 +1,5 @@
 Changelog
 =========
 
+- Functionality for displaying the layers of MEGNet-model migrated into `scripts` directory (`PR2 <https://github.com/mdi-group/gp-net/pull/2>`__)
 - Data download and entry checks migrated into `scripts` directory (`PR1 <https://github.com/mdi-group/gp-net/pull/1>`__)